requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
alembic==1.11.1
pytest==7.4.0
requests==2.31.0
psycopg2-binary==2.9.7
pydantic[email]==1.10.7
fastapi==0.95.2
asyncpg==0.27.0